use a padding character which is part of a hex digit
the _id is going to either be numeric or a mongoid (hex) therefore if we
pad with spaces the batchSlice key will fall outside the expected range
of 00 -> ff.

Of course, the _id could be a different type than numeric or a mongoid -
but I'm uninterested in thinking about that case.